Moved ProfilerAction to standalone module
