+++ composer/profiler/__init__.py
@@ -2,5 +2,5 @@
 
 from composer.profiler.event_handler import ProfilerEventHandler as ProfilerEventHandler
 from composer.profiler.profiler import Profiler as Profiler
-from composer.profiler.profiler import ProfilerAction as ProfilerAction
+from composer.profiler.profiler_action import ProfilerAction as ProfilerAction
 from composer.profiler.profiler_hparams import ProfilerHparams as ProfilerHparams
